+++ src/filesprocessing/filters/SmallerThanFilter.java
@@ -3,17 +3,21 @@ package filesprocessing.filters;
 import java.io.File;
 
 /**
- * Created by Noy on 21-May-17.
- */
-
-/**
  * This class represents a filter which returns files with size strictly smaller than a given
  * number (in K-bytes)
+ * @author Aviad Nissel, Noy Sternlicht
  */
 public class SmallerThanFilter extends Filter {
 
+
+    /* --- Constants --- */
+
     private static final int KBYTES_TO_BYTES = 1024;
-    private int smallerThan;
+
+
+    /* --- Data Members --- */
+
+    private double smallerThan;
 
 
     /* --- Constructors --- */
@@ -25,19 +29,13 @@ public class SmallerThanFilter extends Filter {
      */
     public SmallerThanFilter(String smallerThan, Boolean not){
         super(not);
-        this.smallerThan = (Integer.parseInt(smallerThan)) * KBYTES_TO_BYTES;
+        this.smallerThan = (Double.parseDouble(smallerThan)) * KBYTES_TO_BYTES;
     }
 
-    /* --- Methods --- */
 
+    /* --- Filter Impl. --- */
 
     @Override
-
-    /**
-     * Returns true if the tested file's size is smaller than "smallerThan", false otherwise.
-     * @param file: The tested file object.
-     * @returns: True if the tested file's size is smaller than "smallerThan", false otherwise.
-     */
     public boolean isMatch(File file){
         return file.length() < smallerThan;
     }
